+++ docs/source/api_ref/impl/parser/builtins.rst
@@ -1,0 +1,35 @@
+.. currentmodule:: disnake.ext.components.impl
+
+Builtins Parser Implementations
+===============================
+
+.. automodule:: components.impl.parser.builtins
+
+
+Classes
+-------
+
+.. attributetable:: components.impl.parser.builtins.NoneParser
+
+.. autoclass:: components.impl.parser.builtins.NoneParser
+    :members:
+
+.. attributetable:: components.impl.parser.builtins.FloatParser
+
+.. autoclass:: components.impl.parser.builtins.FloatParser
+    :members:
+
+.. attributetable:: components.impl.parser.builtins.IntParser
+
+.. autoclass:: components.impl.parser.builtins.IntParser
+    :members:
+
+.. attributetable:: components.impl.parser.builtins.BoolParser
+
+.. autoclass:: components.impl.parser.builtins.BoolParser
+    :members:
+
+.. attributetable:: components.impl.parser.builtins.StringParser
+
+.. autoclass:: components.impl.parser.builtins.StringParser
+    :members:
