+++ docs/source/api_ref/impl/parser/index.rst
@@ -10,4 +10,5 @@ Submodules
    :maxdepth: 1
 
    base </api_ref/impl/parser/base>
-   stdlib </api_ref/impl/parser/stdlib>
+   builtins </api_ref/impl/parser/builtins>
+   datetime </api_ref/impl/parser/datetime>
