--- docs/source/api_ref/impl/parser/stdlib.rst
@@ -1,66 +1,0 @@
-.. currentmodule:: disnake.ext.components.impl
-
-Standard Library Parser Implementation
-======================================
-
-.. automodule:: components.impl.parser.stdlib
-
-
-Enumerations
-------------
-
-.. autoenum:: components.impl.parser.stdlib.Resolution
-    :members:
-
-Classes
--------
-
-.. attributetable:: components.impl.parser.stdlib.NoneParser
-
-.. autoclass:: components.impl.parser.stdlib.NoneParser
-    :members:
-
-.. attributetable:: components.impl.parser.stdlib.FloatParser
-
-.. autoclass:: components.impl.parser.stdlib.FloatParser
-    :members:
-
-.. attributetable:: components.impl.parser.stdlib.IntParser
-
-.. autoclass:: components.impl.parser.stdlib.IntParser
-    :members:
-
-.. attributetable:: components.impl.parser.stdlib.BoolParser
-
-.. autoclass:: components.impl.parser.stdlib.BoolParser
-    :members:
-
-.. attributetable:: components.impl.parser.stdlib.StringParser
-
-.. autoclass:: components.impl.parser.stdlib.StringParser
-    :members:
-
-.. attributetable:: components.impl.parser.stdlib.DatetimeParser
-
-.. autoclass:: components.impl.parser.stdlib.DatetimeParser
-    :members:
-
-.. attributetable:: components.impl.parser.stdlib.TimedeltaParser
-
-.. autoclass:: components.impl.parser.stdlib.TimedeltaParser
-    :members:
-
-.. attributetable:: components.impl.parser.stdlib.DateParser
-
-.. autoclass:: components.impl.parser.stdlib.DateParser
-    :members:
-
-.. attributetable:: components.impl.parser.stdlib.TimeParser
-
-.. autoclass:: components.impl.parser.stdlib.TimeParser
-    :members:
-
-.. attributetable:: components.impl.parser.stdlib.TimezoneParser
-
-.. autoclass:: components.impl.parser.stdlib.TimezoneParser
-    :members:
